# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KaraoKê Queue API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  OPEN: 'OPEN',
  PAUSED: 'PAUSED',
  CLOSED: 'CLOSED',
} as const;

export interface Session {
  id: string;
  name: string;
  status: SessionStatus;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  /** @minLength 1 */
  name: string;
}

export type SessionStatusUpdateStatus = typeof SessionStatusUpdateStatus[keyof typeof SessionStatusUpdateStatus];


export const SessionStatusUpdateStatus = {
  OPEN: 'OPEN',
  PAUSED: 'PAUSED',
  CLOSED: 'CLOSED',
} as const;

export interface SessionStatusUpdate {
  status: SessionStatusUpdateStatus;
}

export interface SessionSummary {
  sessionId: string;
  totalReservations: number;
  totalParticipants: number;
  completed: number;
  skipped: number;
  removed: number;
  queued: number;
}

export interface Participant {
  id: string;
  cpf: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface ParticipantInput {
  /**
     * @minLength 11
     * @maxLength 11
     */
  cpf: string;
  /** @minLength 1 */
  name: string;
}

export interface SongSearchResult {
  youtubeId: string;
  title: string;
  channelName: string;
  thumbnailUrl: string;
}

export interface Song {
  id: string;
  youtubeId: string;
  title: string;
  channelName: string;
  thumbnailUrl: string;
  embeddable: boolean;
  cachedAt: string;
  updatedAt: string;
}

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];


export const ReservationStatus = {
  QUEUED: 'QUEUED',
  PLAYING: 'PLAYING',
  FINISHED: 'FINISHED',
  SKIPPED: 'SKIPPED',
  REMOVED: 'REMOVED',
} as const;

export interface Reservation {
  id: string;
  sessionId: string;
  participantId: string;
  songId: string;
  status: ReservationStatus;
  requestedAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  skippedAt?: string | null;
  /** @nullable */
  removedAt?: string | null;
  participant: Participant;
  song: Song;
}

export interface QueueEntry {
  id: string;
  sessionId: string;
  reservationId: string;
  position: number;
  addedAt: string;
  reservation: Reservation;
}

export interface ReservationInput {
  /**
     * @minLength 11
     * @maxLength 11
     */
  cpf: string;
  /** @minLength 1 */
  name: string;
  youtubeId: string;
  title: string;
  channelName: string;
  thumbnailUrl: string;
}

export interface OperatorCredentials {
  username: string;
  password: string;
}

export interface OperatorSession {
  id: string;
  username: string;
  displayName: string;
}

export type SearchSongsParams = {
q: string;
};