# lib/api-zod/src/generated/types/operatorSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KaraoKê Queue API
 * OpenAPI spec version: 0.1.0
 */

export interface OperatorSession {
  id: string;
  username: string;
  displayName: string;
}